+++ android/src/main/kotlin/com/example/flutter_mongo_stitch/MyMongoStitchClient.kt
@@ -88,7 +88,7 @@ class MyMongoStitchClient(
     }
 
     fun signInWithGoogle(idToken: String, callback: App.Callback<User>): RealmAsyncTask? {
-        return app.loginAsync(Credentials.google(idToken ,GoogleAuthType.ID_TOKEN), callback)
+        return app.loginAsync(Credentials.google(idToken ,GoogleAuthType.AUTH_CODE), callback)
     }
 
     fun signInWithFacebook(accessToken: String, callback: App.Callback<User>): RealmAsyncTask? {
@@ -107,6 +107,15 @@ class MyMongoStitchClient(
 
     fun signInWithApple(idToken: String, callback: App.Callback<User>) : RealmAsyncTask? {
         return app.loginAsync(Credentials.apple(idToken), callback);
+    }
+
+    fun isLoggedIn() : Boolean {
+        return app.currentUser()?.isLoggedIn ?: false
+    }
+
+    fun linkCredentials(credsJson: Map<String, Any>, callback: App.Callback<User>) : RealmAsyncTask? {
+        val creds = CredentialsExtensions.fromMap(credsJson)
+        return app.currentUser()?.linkCredentialsAsync(creds, callback)
     }
 
     fun logout(callback: App.Callback<User>): RealmAsyncTask?
@@ -134,7 +143,6 @@ class MyMongoStitchClient(
             : RealmResultTask<InsertOneResult>? {
         val collection = getCollection(databaseName, collectionName)
 
-        
         //Document.parse(json)
         val document = Document()
 
@@ -345,6 +353,4 @@ class MyMongoStitchClient(
             this.client = user.getMongoClient("mongodb-atlas")
         }
     }
-
-
 }