+++ example/web/stitchUtils.js
@@ -2,11 +2,28 @@
 
 var mongoClient;
 var stitchAppClient;
+var realmApp;
 
 function uint8ArrayToHex(uint8Array) {
     return Array.prototype.map.call(new Uint8Array(uint8Array.buffer), x => ('00' + x.toString(16)).slice(-2)).join('');
 }
 
+function getCredFromJson(json){
+    switch(json["type"]){
+        case "anon":
+            throw Exception("can't link anonymous")
+        case "email_password":
+            return Realm.Credentials.emailPassword(json["email"] as String, json["password"] as String)
+        case "apple":
+            return Realm.Credentials.apple(json["idToken"] as String)
+        case "facebook":
+            return Realm.Credentials.facebook(json["accessToken"] as String)
+//        case "google":
+//            return Credentials.google(json["authorizationCode"] as String)
+        case "jwt" :
+            return Realm.Credentials.jwt(json["jwtToken"] as String)
+    }
+}
 
 function Mongo() {
     Mongo.prototype.connectMongo  = function(appId) {
@@ -18,6 +35,9 @@ function Mongo() {
         );
 
         this.sendAuthListenerEvent(null);
+
+        // Realm
+        realmApp = new Realm.App({ id: appId });
     }
 
     /// -----------------------------------------------------
@@ -82,6 +102,13 @@ function Mongo() {
 
         var strings = [];
         results.forEach((doc) => {
+
+            let docId = doc['_id']
+            if (typeof docId === 'object' || docId instanceof Object){
+                doc['_id'] = {
+                    '$oid': `${docId}`
+                }
+            }
             strings.push(JSON.stringify(doc))
         })
 
@@ -270,6 +297,13 @@ function Mongo() {
         console.log('DONE!');
     };
 
+    Mongo.prototype.linkCredentials = async function(jsonData){
+         var json = JSON.parse(jsonData);
+        const realmUserCredentials = getCredFromJson(json)
+        await realmApp.currentUser.linkCredentials(realmUserCredentials);
+        return user;
+    }
+
     Mongo.prototype.logout  = async function(){
         await stitchAppClient.auth.logout();
         this.sendAuthListenerEvent(null);
@@ -329,33 +363,46 @@ function Mongo() {
      Mongo.prototype.setupWatchCollection = async function(databaseName, collectionName, arg){
         var collection = this.getCollection(databaseName, collectionName)
 
-        console.log(arg)
+
         console.log(typeof arg)
-        if (typeof arg === 'string' || arg instanceof String){
+        let asObjectIDs = false;
+        if (arg == null){
+            asObjectIDs = true;
+        }
+        else if (typeof arg === 'string' || arg instanceof String){
             arg = JSON.parse(arg);
         }
-        if (typeof arg === 'array' || arg instanceof Array){
+        else if (typeof arg === 'array' || arg instanceof Array){
             if(arg[1] == false){
                 arg = arg[0]
             }
             else {
-                var lst = [];
-                arg[0].forEach((str) => {
-                    lst.push(new stitch.BSON.ObjectId(str))
-                })
-                arg = lst;
+                asObjectIDs = true
+                if(arg[0] == null){
+                    arg = null
+                }
+                else{
+                    var lst = [];
+                    arg[0].forEach((str) => {
+                        lst.push(new stitch.BSON.ObjectId(str))
+                    })
+                    arg = lst;
+                }
             }
         }
 
-
         var changeStream = await collection.watch(arg);
 
+        console.log(`asObjectIDs=${asObjectIDs}`)
         // Set the change listener. This will be called
         // when the watched documents are updated.
         changeStream.onNext((event) => {
 
+          let docId = event.fullDocument['_id']
           var results = {
-            "_id": event.fullDocument['_id']
+            "_id": asObjectIDs
+                ? {'$oid': docId} //new stitch.BSON.ObjectId(docId)
+                : docId
           }
           var watchEvent = new CustomEvent("watchEvent."+databaseName+"."+collectionName, {
                detail: JSON.stringify(results)