+++ ios/Classes/extentions.swift
@@ -50,3 +50,30 @@ extension User{
         ]
     }
 }
+
+class CredentialsExtensions{
+    static func fromMap(_ json: Dictionary<String, Any>) throws -> Credentials?{
+        enum MyError: Error {
+              case cantLink
+          }
+        
+        let type = json["type"] as! String
+        switch(type){
+        case "anon" :
+            throw MyError.cantLink
+        case "email_password":
+            return Credentials.emailPassword(email: json["email"] as! String, password: json["password"] as! String)
+        case "apple":
+            return Credentials.apple(idToken: json["idToken"] as! String)
+        case "facebook":
+            return Credentials.facebook(accessToken: json["accessToken"] as! String)
+        //case "google" : return Credentials.google(json["authorizationCode"] as String)
+        case "jwt":
+            return Credentials.jwt(token: json["jwtToken"] as! String)
+        default:
+            break
+        }
+        return nil
+    }
+    
+}