+++ lib/auth/credentials/custom_jwt_credential.dart
@@ -5,4 +5,12 @@ class CustomJwtCredential extends StitchCredential {
   final String token;
 
   CustomJwtCredential(this.token);
+
+  @override
+  Map<String, Object> toJson() {
+    return {
+      "type": "jwt",
+      "token": token
+    };
+  }
 }