+++ lib/auth/credentials/google_credential.dart
@@ -1,17 +1,15 @@
-import 'package:flutter/cupertino.dart';
-
 import 'stitch_credential.dart';
 
-@deprecated
-class GoogleCredential extends StitchCredential {
-  final List<String> scopes;
-  final String serverClientId;
-
-  GoogleCredential({@required this.serverClientId, this.scopes});
-}
-
 class GoogleCredential2 extends StitchCredential {
-  final String accessToken;
+  final String authCode;
 
-  GoogleCredential2(this.accessToken);
+  GoogleCredential2(this.authCode);
+
+  @override
+  Map<String, Object> toJson() {
+    return {
+      "type": "google",
+      "authCode": authCode
+    };
+  }
 }