+++ lib/auth/credentials/user_password_credentinal.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/foundation.dart';
-
 import 'stitch_credential.dart';
 
 @deprecated
@@ -8,7 +6,16 @@ class UserPasswordCredential extends StitchCredential {
   final String password;
 
   UserPasswordCredential({
-    @required this.username,
-    @required this.password,
+    required this.username,
+    required this.password,
   });
+
+  @override
+  Map<String, Object> toJson() {
+    return {
+      "type": "email_password",
+      "email": username,
+      "password": password,
+    };
+  }
 }