+++ lib/auth/mongo_realm_auth.dart
@@ -1,11 +1,9 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter_mongodb_realm/auth/credentials/apple_credential.dart';
 
 //import 'package:flutter_facebook_login/flutter_facebook_login.dart';
-import 'package:flutter_mongodb_realm/auth/credentials/google_credential.dart';
 import 'package:flutter_mongodb_realm/database/mongo_document.dart';
 import 'package:flutter_mongodb_realm/google_sign_in_git_mock/google_sign_in.dart';
 // import 'package:flutter_mongodb_realm/google_sign_in_git_mock/google_sign_in.dart';
@@ -31,7 +29,8 @@ class MongoRealmAuth {
   /// Logs in as a user with the given credentials associated with an
   /// authentication provider.
   @deprecated
-  Future<CoreRealmUser> loginWithCredential(StitchCredential credential) async {
+  Future<CoreRealmUser?> loginWithCredential(
+      StitchCredential credential) async {
     var result;
 
     if (credential is AnonymousCredential) {
@@ -43,22 +42,7 @@ class MongoRealmAuth {
       );
     } else if (credential is GoogleCredential2) {
       try {
-        var accessToken = credential.accessToken;
-        result = await FlutterMongoRealm.signInWithGoogle(accessToken);
-      } on Exception catch (e) {
-        print(e);
-      }
-    } else if (credential is GoogleCredential) {
-      _googleLoginWrapper.init(
-        serverClientId:
-            "${credential.serverClientId}.apps.googleusercontent.com",
-        scopes: credential.scopes,
-      );
-
-      try {
-        var authCode =
-            await _googleLoginWrapper.handleSignInAndGetAuthServerCode();
-        print(authCode ?? 'nothing');
+        var authCode = credential.authCode;
         result = await FlutterMongoRealm.signInWithGoogle(authCode);
       } on Exception catch (e) {
         print(e);
@@ -81,37 +65,47 @@ class MongoRealmAuth {
     return result;
   }
 
-  Future<bool> logout() async {
+  Future<CoreRealmUser?> linkCredentials(StitchCredential credential) async{
+    var result = await FlutterMongoRealm.linkCredentials(credential.toJson());
+    return result;
+  }
+
+  Future<bool> isLoggedIn() async {
+    var result = await FlutterMongoRealm.isLoggedIn();
+    return result;
+  }
+
+  Future<bool?> logout() async {
     var result = await FlutterMongoRealm.logout();
 
-    bool loggedWithGoogle = await _googleLoginWrapper.isLogged;
-//    bool loggedWithFacebook = await _facebookLoginWrapper.isLogged;
-//
+    bool loggedWithGoogle =
+        await (_googleLoginWrapper.isLogged as FutureOr<bool>);
     if (loggedWithGoogle) await _googleLoginWrapper.handleSignOut();
-//
+
+//    bool loggedWithFacebook = await _facebookLoginWrapper.isLogged;
 //    if (loggedWithFacebook)
 //      await _facebookLoginWrapper.handleSignOut();
 
     return result;
   }
 
-  Future<String> getUserId() async {
+  Future<String?> getUserId() async {
     var result = await FlutterMongoRealm.getUserId();
     return result;
   }
 
   Future<bool> registerWithEmail(
-      {@required String email, @required String password}) async {
+      {required String email, required String password}) async {
     var result = await FlutterMongoRealm.registerWithEmail(email, password);
     return result;
   }
 
-  Future<bool> sendResetPasswordEmail(String email) async {
+  Future<bool> sendResetPasswordEmail(String? email) async {
     var result = await FlutterMongoRealm.sendResetPasswordEmail(email);
     return result;
   }
 
-  Future<CoreRealmUser> get user async => await FlutterMongoRealm.getUser();
+  Future<CoreRealmUser?> get user async => await FlutterMongoRealm.getUser();
 
   Stream authListener() {
     var stream = FlutterMongoRealm.authListener();
@@ -122,22 +116,22 @@ class MongoRealmAuth {
 /// ////////////////////////////////////////////////////////////////
 
 class _GoogleLoginWrapper {
-  GoogleSignIn _googleSignIn;
+  GoogleSignIn? _googleSignIn;
 
-  Future<bool> get isLogged =>
-      _googleSignIn == null ? Future.value(false) : _googleSignIn.isSignedIn();
+  Future<bool?> get isLogged =>
+      _googleSignIn == null ? Future.value(false) : _googleSignIn!.isSignedIn();
 
-  init({@required String serverClientId, List<String> scopes}) {
+  init({required String serverClientId, List<String>? scopes}) {
     _googleSignIn =
         GoogleSignIn(serverClientId: serverClientId, scopes: scopes);
   }
 
-  Future<String> handleSignInAndGetAuthServerCode() async {
+  Future<String?> handleSignInAndGetAuthServerCode() async {
     assert(_googleSignIn != null);
 
-    String code;
+    String? code;
     // try {
-    var account = await _googleSignIn.signIn();
+    var account = await _googleSignIn!.signIn();
 
     if (account != null) code = account.serverAuthCode;
     // } on Exception catch (error) {
@@ -147,5 +141,5 @@ class _GoogleLoginWrapper {
     return code;
   }
 
-  Future<void> handleSignOut() => _googleSignIn.disconnect();
+  Future<void> handleSignOut() => _googleSignIn!.disconnect();
 }