+++ lib/bson_document.dart
@@ -2,16 +2,16 @@ import 'dart:convert';
 import 'package:bson/bson.dart' show ObjectId;
 
 /// Extensions for ObjectId
-extension ObjectIdExtentions on ObjectId {
+extension ObjectIdExtentions on ObjectId? {
   /// Correct supported json format
-  toJsonOid() => {"\$oid": "${this.toHexString()}"};
+  toJsonOid() => {"\$oid": "${this!.toHexString()}"};
 }
 
 // TODO: maybe using 'BsonObject' instead 'dynamic'
 class BsonDocument {
-  final Map<String, dynamic> _map;
+  final Map<String, dynamic>? _map;
 
-  Map<String, dynamic> get map => _map;
+  Map<String, dynamic>? get map => _map;
 
   /// Creates a document instance initialized with the given map.
   /// or an empty Document instance if not provided.
@@ -23,7 +23,7 @@ class BsonDocument {
     }
 
     var __map = <String, dynamic>{};
-    for (var entry in _map.entries) {
+    for (var entry in _map!.entries) {
       if (entry.value is ObjectId)
         __map[entry.key] = (entry.value as ObjectId).toJsonOid();
       else if (entry.value is Map) {
@@ -46,7 +46,7 @@ class BsonDocument {
                 for (int index = 0; index < list.length; index++) {
                   if (list[index] is ObjectId) {
                     __map[entry.key][entry2.key][entry3.key][index] =
-                        (list[index].value as ObjectId).toJsonOid();
+                        (list[index].value as ObjectId?)!.toJsonOid();
                   } else {
                     __map[entry.key][entry2.key][entry3.key][index] =
                         list[index];